Splits smoke tests
